+++ src/main/java/com/guxingke/intellij/plugin/postfix/ToolsPostfixTemplateProvider.java
@@ -1,5 +1,6 @@
 package com.guxingke.intellij.plugin.postfix;
 
+import com.guxingke.intellij.plugin.Configs;
 import com.guxingke.intellij.plugin.postfix.template.TemplateFactory;
 import com.guxingke.intellij.plugin.postfix.template.struct.StructMapperPostfixTemplate;
 import com.intellij.codeInsight.template.postfix.templates.PostfixTemplate;
@@ -7,6 +8,7 @@ import com.intellij.codeInsight.template.postfix.templates.PostfixTemplateProvid
 import com.intellij.openapi.editor.Editor;
 import com.intellij.psi.PsiFile;
 import java.util.Collection;
+import java.util.HashSet;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -17,10 +19,38 @@ public class ToolsPostfixTemplateProvider implements PostfixTemplateProvider {
   private final Set<PostfixTemplate> templates;
 
   public ToolsPostfixTemplateProvider() {
-    var internals = Set.of(new StructMapperPostfixTemplate(this));
+    var cfg = Configs.getConfig();
+    if (!cfg.getPostfix().isEnable()) {
+      templates = new HashSet<>();
+      return;
+    }
 
-    var builtinTemplates = TemplateFactory.createBuiltinTemplates(this);
-    var externals = TemplateFactory.createTemplates(this);
+    var ic = cfg.getPostfix().getInternal();
+    Set<PostfixTemplate> internals = new HashSet<>();
+    if (ic.isEnable()) {
+      internals = Set.of(new StructMapperPostfixTemplate(this))
+          .stream()
+          .filter(it -> !ic.getBlocklist().contains(it.getPresentableName()))
+          .collect(Collectors.toSet());
+    }
+
+    Set<PostfixTemplate> builtinTemplates = new HashSet<>();
+    var bc = cfg.getPostfix().getBuiltin();
+    if (bc.isEnable()) {
+      builtinTemplates = TemplateFactory.createBuiltinTemplates(this)
+          .stream()
+          .filter(it -> !bc.getBlocklist().contains(it.getPresentableName()))
+          .collect(Collectors.toSet());
+    }
+
+    Set<PostfixTemplate> externals = new HashSet<>();
+    var ec = cfg.getPostfix().getExternal();
+    if (ec.isEnable()) {
+      externals = TemplateFactory.createExternalTemplates(this)
+          .stream()
+          .filter(it -> !ec.getBlocklist().contains(it.getPresentableName()))
+          .collect(Collectors.toSet());
+    }
     templates = Stream.of(externals, builtinTemplates, internals)
         .flatMap(Collection::stream)
         .collect(Collectors.toSet());